added delete button and functionality
