+++ api/database.json
@@ -20,13 +20,6 @@
       "title": "JavaScript Promises",
       "body": "Explored promises, asynchronous programming, and how they work with async/await in JS.",
       "moodId": "3"
-    },
-    {
-      "id": "1f14",
-      "date": "2024-12-07",
-      "title": "test",
-      "body": "r",
-      "moodId": "3"
     }
   ],
   "moods": [
