+++ scripts/database.js
@@ -1,3 +1,5 @@
+import { transientState } from "./transient-state.js";
+
 export const getJournalEntries = () => {
   return fetch("http://localhost:8088/entries?_embed=mood") // Fetch from the API
     .then((res) => res.json()) // Parse as JSON
@@ -23,6 +25,30 @@ export const saveJournalEntry = async (newEntry) => {
       const event = new CustomEvent("entrySaved");
       window.dispatchEvent(event);
     });
+};
+
+export const postEntry = async () => {
+  const postOptions = {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(transientState),
+  };
+
+  const res = await fetch("http://localhost:8088/entries", postOptions);
+
+  const newEntryEvent = new CustomEvent("newEntry");
+  document.dispatchEvent(newEntryEvent);
+};
+
+export const deleteEntry = async (entryId) => {
+  const res = await fetch(`http://localhost:8088/entries/${entryId}`, {
+    method: "DELETE",
+  });
+
+  const newDeleteEvent = new CustomEvent("deletedEntry");
+  document.dispatchEvent(newDeleteEvent);
 };
 
 export const getMoods = () => {
