+++ scripts/transient-state.js
@@ -1,4 +1,4 @@
-const transientState = {
+export const transientState = {
   date: "",
   title: "",
   body: "",
@@ -20,19 +20,4 @@ export const setBody = (chosenBody) => {
 export const setMood = (chosenMood) => {
   transientState.moodId = chosenMood;
   console.log(transientState);
-};
-
-export const postEntry = async () => {
-  const postOptions = {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(transientState),
-  };
-
-  const res = await fetch("http://localhost:8088/entries", postOptions);
-
-  const newEntryEvent = new CustomEvent("newEntry");
-  document.dispatchEvent(newEntryEvent);
 };
